state.c: fix warning
state.c: In function ‘clientid_free’:
state.c:641: warning: cast to pointer from integer of different size

+++ state.c
@@ -638,7 +638,8 @@ static void clientid_free(struct nfs_clientid *clid)
 	else
 		evtimer_del(&clid->timer);
 
-	g_hash_table_remove(srv.clid_idx, (void *) clid->id_short);
+	g_hash_table_remove(srv.clid_idx,
+			    (void *)(unsigned long) clid->id_short);
 
 	free(clid->id.buf);
 	free_cb_client4(&clid->callback);
